[GREEN] Finish separating register page into multiple components and passed the test.
